Merge development into ngwpc-candidate for release 3.1.2.4.0-rc1

+++ src/troute-network/troute/network/reservoirs/hybrid/hybrid_structs.h
@@ -3,13 +3,14 @@
 /*
     C Structures
 */
+#include <stdint.h>
 #include "../../reach_structs.h"
 
 #define HYBRID_MAX_PATH_LENGTH 256
 #define HYBRID_MAX_START_DATE_LENGTH 20 
 
 typedef struct {
-  int lake_number;
+  int64_t lake_number;
   float dam_length, area, max_depth;
   float orifice_area, orifice_coefficient, orifice_elevation;
   float weir_coefficient, weir_elevation, weir_length;
